moved `detect_release` method to `Util::Release`
