+++ lib/raven/configuration.rb
@@ -198,7 +198,7 @@ module Raven
       self.project_root = detect_project_root
       self.rails_activesupport_breadcrumbs = false
       self.rails_report_rescued_exceptions = true
-      self.release = detect_release
+      self.release = Utils::Release.detect_release
       self.sample_rate = 1.0
       self.sanitize_credit_cards = true
       self.sanitize_fields = []
@@ -319,14 +319,6 @@ module Raven
       end
     end
 
-    def detect_release
-      detect_release_from_git ||
-        detect_release_from_capistrano ||
-        detect_release_from_heroku
-    rescue => ex
-      logger.error "Error detecting release: #{ex.message}"
-    end
-
     def excluded_exception?(exc)
       excluded_exceptions.any? { |x| get_exception_class(x) === exc }
     end
@@ -344,41 +336,6 @@ module Raven
         x.split(MODULE_SEPARATOR).reject(&:empty?).inject(Object) { |a, e| a.const_get(e) }
       end
     rescue NameError # There's no way to safely ask if a constant exist for an unknown string
-      nil
-    end
-
-    def detect_release_from_heroku
-      return unless running_on_heroku?
-      logger.warn(heroku_dyno_metadata_message) && return unless ENV['HEROKU_SLUG_COMMIT']
-
-      ENV['HEROKU_SLUG_COMMIT']
-    end
-
-    def running_on_heroku?
-      File.directory?("/etc/heroku")
-    end
-
-    def heroku_dyno_metadata_message
-      "You are running on Heroku but haven't enabled Dyno Metadata. For Sentry's "\
-      "release detection to work correctly, please run `heroku labs:enable runtime-dyno-metadata`"
-    end
-
-    def detect_release_from_capistrano
-      revision_file = File.join(project_root, 'REVISION')
-      revision_log = File.join(project_root, '..', 'revisions.log')
-
-      if File.exist?(revision_file)
-        File.read(revision_file).strip
-      elsif File.exist?(revision_log)
-        File.open(revision_log).to_a.last.strip.sub(/.*as release ([0-9]+).*/, '\1')
-      end
-    end
-
-    def detect_release_from_git
-      release = `git rev-parse --short HEAD 2>/dev/null`.strip if File.directory?('.git')
-      return nil if release == ''
-      release
-    rescue StandardError
       nil
     end
 
@@ -433,6 +390,10 @@ module Raven
       else
         resolve_hostname
       end
+    end
+
+    def running_on_heroku?
+      File.directory?("/etc/heroku")
     end
   end
 end
